crypto: Fix CHERI issues
Fix cheri-related warnings from the compiler and from clang-tidy.

- Use uintptr_t for the data members in parser contexts.
- Use the macros to convert pointers to integers where needed.

Co-developed-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>

+++ crypto/scompress.c
@@ -29,7 +29,7 @@ struct scomp_scratch {
 	spinlock_t	lock;
 	union {
 		void	*src;
-		unsigned long saddr;
+		uintptr_t saddr;
 	};
 };
